Implement retry logic in cache
The retry logic is still implemented with recursion since
I cannot figure out a way to write asynchronous recursion
in iterative style with for-loop.

It may be possible to implement with for-loop given that
the function is defined wht the keyword `async` from ES7
so that I can `yield` a async operation in a for-loop, making
the for-loop async.